factor out export base
